Move `authenticateIfNeed` method to `SendBirdCall & Create extension files for `SendBirdCall` and `UIApplication`
