+++ QuickStart/AppDelegate+Intent.swift
@@ -19,26 +19,14 @@ extension AppDelegate {
     func application(_ app: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
         guard let calleeId = url.calleeId else { return false }
         
-        guard SendBirdCall.currentUser == nil else {
-            self.dial(to: calleeId, hasVideo: false)
-            return true
-        }
-        
-        let userId = UserDefaults.standard.user.id
-        let accessToken = UserDefaults.standard.accessToken
-        if userId.isEmpty { return false }
-        
-        let authParams = AuthenticateParams(userId: userId, accessToken: accessToken)
-        SendBirdCall.authenticate(with: authParams) { [weak self] user, error in
-            guard let self = self else { return }
+        SendBirdCall.authenticateIfNeed { error in
             guard error == nil else {
-                DispatchQueue.main.async { [ weak self] in
-                    guard let self = self else { return }
-                    self.showError(with: error?.localizedDescription ?? "Failed to call with unknown error")
+                DispatchQueue.main.async {
+                    UIApplication.shared.showError(with: error?.localizedDescription ?? "Failed to call with unknown error")
                 }
                 return
             }
-            self.dial(to: calleeId, hasVideo: false)
+            SendBirdCall.dialFromExternal(to: calleeId, hasVideo: false)
         }
         return true
     }
@@ -49,82 +37,25 @@ extension AppDelegate {
     
     func application(_ application: UIApplication, continue userActivity: NSUserActivity, restorationHandler: @escaping ([UIUserActivityRestoring]?) -> Void) -> Bool {
         guard let calleeId = userActivity.calleeId else {
-            showError(with: "Could not determine callee ID")
+            UIApplication.shared.showError(with: "Could not determine callee ID")
             return false
         }
         
         guard let hasVideo = userActivity.hasVideo else {
-            showError(with: "Could not determine video from call log")
+            UIApplication.shared.showError(with: "Could not determine video from call log")
             return false
         }
         
-        guard SendBirdCall.currentUser == nil else {
-            self.dial(to: calleeId, hasVideo: hasVideo)
-            return true
-        }
-        
-        let userId = UserDefaults.standard.user.id
-        let accessToken = UserDefaults.standard.accessToken
-        guard !userId.isEmpty else { return false }
-        
-        let authParams = AuthenticateParams(userId: userId, accessToken: accessToken)
-        SendBirdCall.authenticate(with: authParams) { [weak self] user, error in
-            guard let self = self else { return }
+        SendBirdCall.authenticateIfNeed { error in
             guard error == nil else {
-                DispatchQueue.main.async { [ weak self] in
-                    guard let self = self else { return }
-                    self.showError(with: error?.localizedDescription ?? "Failed to call with unknown error")
+                DispatchQueue.main.async {
+                    UIApplication.shared.showError(with: error?.localizedDescription ?? "Failed to call with unknown error")
                 }
                 return
             }
-            self.dial(to: calleeId, hasVideo: hasVideo)
-        }
-        return true
-    }
-    
-    // MARK: - Actions
-    
-    private func dial(to calleeId: String, hasVideo video: Bool) {
-        let callOption = CallOptions(isAudioEnabled: true, isVideoEnabled: video, localVideoView: nil, remoteVideoView: nil, useFrontCamera: true)
-        let dialParams = DialParams(calleeId: calleeId, isVideoCall: video, callOptions: callOption, customItems: [:])
-        SendBirdCall.dial(with: dialParams) { call, error in
-            guard let call = call, error == nil else {
-                DispatchQueue.main.async { [ weak self] in
-                    guard let self = self else { return }
-                    self.showError(with: error?.localizedDescription ?? "Failed to call with unknown error")
-                }
-                return
-            }
-            DispatchQueue.main.async { [weak self] in
-                guard let self = self else { return }
-                self.showCallView(call, hasVideo: call.isVideoCall)
-            }
-        }
-    }
-    
-    private func showCallView(_ call: DirectCall, hasVideo: Bool) {
-        let storyboard = UIStoryboard.init(name: "Main", bundle: nil)
-        let viewController = storyboard.instantiateViewController(withIdentifier: call.isVideoCall ? "VideoCallViewController" : "VoiceCallViewController")
-
-        if var dataSource = viewController as? DirectCallDataSource {
-            dataSource.call = call
-            dataSource.isDialing = true
+            SendBirdCall.dialFromExternal(to: calleeId, hasVideo: hasVideo)
         }
         
-        if let topViewController = UIViewController.topViewController {
-            topViewController.present(viewController, animated: true, completion: nil)
-        } else {
-            UIApplication.shared.keyWindow?.rootViewController = viewController
-            UIApplication.shared.keyWindow?.makeKeyAndVisible()
-        }
-    }
-    
-    private func showError(with message: String) {
-        if let topViewController = UIViewController.topViewController {
-            topViewController.presentErrorAlert(message: message)
-        } else {
-            UIApplication.shared.keyWindow?.rootViewController?.presentErrorAlert(message: message)
-            UIApplication.shared.keyWindow?.makeKeyAndVisible()
-        }
+        return true
     }
 }
